+++ telegram-longpollingbot-springboot-starter/src/main/java/org/telegram/telegrambot/aspect/ExceptionHandlerAspect.java
@@ -1,4 +1,4 @@
-package org.telegram.telegrambot.aop;
+package org.telegram.telegrambot.aspect;
 
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
@@ -7,12 +7,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambot.container.ExceptionMappingMethodContainer;
-import org.telegram.telegrambot.dto.InvocationUnit;
-import org.telegram.telegrambot.dto.MethodTargetPair;
+import org.telegram.telegrambot.model.InvocationUnit;
+import org.telegram.telegrambot.model.MethodTargetPair;
 import org.telegram.telegrambot.handler.ApiMethodsReturningMethodInvoker;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
-import java.lang.reflect.Method;
 import java.util.Optional;
 
 @Aspect
