+++ telegram-longpollingbot-springboot-starter/src/main/java/org/telegram/telegrambot/container/UpdateMappingMethodContainer.java
@@ -1,7 +1,7 @@
 package org.telegram.telegrambot.container;
 
 import org.springframework.stereotype.Component;
-import org.telegram.telegrambot.dto.MethodTargetPair;
+import org.telegram.telegrambot.model.MethodTargetPair;
 
 import java.util.List;
