+++ telegram-longpollingbot-springboot-starter/src/main/java/org/telegram/telegrambot/handler/ApiMethodsReturningMethodInvoker.java
@@ -1,7 +1,6 @@
 package org.telegram.telegrambot.handler;
 
-import org.telegram.telegrambot.dto.InvocationUnit;
-import org.telegram.telegrambot.dto.MethodTargetPair;
+import org.telegram.telegrambot.model.InvocationUnit;
 import org.telegram.telegrambots.meta.api.methods.PartialBotApiMethod;
 import org.telegram.telegrambots.meta.api.objects.Message;
