+++ telegram-longpollingbot-springboot-starter/src/main/java/org/telegram/telegrambot/handler/ApiMethodsReturningMethodInvokerImpl.java
@@ -4,7 +4,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import org.springframework.util.ReflectionUtils;
-import org.telegram.telegrambot.dto.InvocationUnit;
+import org.telegram.telegrambot.model.InvocationUnit;
 import org.telegram.telegrambots.meta.api.methods.PartialBotApiMethod;
 import org.telegram.telegrambots.meta.api.objects.Message;
