+++ telegram-longpollingbot-springboot-starter/src/main/java/org/telegram/telegrambot/handler/UpdateMappingMethodProviderImpl.java
@@ -9,8 +9,8 @@ import org.telegram.telegrambot.annotation.UpdateMapping;
 import org.telegram.telegrambot.container.StringToObjectMapperContainer;
 import org.telegram.telegrambot.container.UpdateMappingMethodContainer;
 import org.telegram.telegrambot.databind.StringToObjectMapper;
-import org.telegram.telegrambot.dto.InvocationUnit;
-import org.telegram.telegrambot.dto.MethodTargetPair;
+import org.telegram.telegrambot.model.InvocationUnit;
+import org.telegram.telegrambot.model.MethodTargetPair;
 import org.telegram.telegrambot.expection.StringToObjectMapperException;
 import org.telegram.telegrambot.repository.StateSource;
 import org.telegram.telegrambots.meta.api.objects.Update;
