+++ telegram-longpollingbot-springboot-starter/src/main/java/org/telegram/telegrambot/model/InvocationUnit.java
@@ -1,4 +1,4 @@
-package org.telegram.telegrambot.dto;
+package org.telegram.telegrambot.model;
 
 import java.lang.reflect.Method;
